Add files via upload
Text Sentiment Analyzer
A Python-based sentiment analysis tool that classifies the sentiment of text data as Positive, Negative, or Neutral. This project utilizes TextBlob for natural language processing to analyze user input or text from predefined datasets, making it an easy-to-use solution for sentiment analysis. The results are visualized using Matplotlib and Seaborn to showcase the distribution of sentiments.